JCBC-694: Add method to prepare String Statement
Motivation
----------
While creating Query from a simple String is possible in the API, no
public way of doing the same when preparing a QueryPlan exist.

Modifications
-------------
Added an override on Bucket and AsyncBucket to accept preparation of
String statements.

This is using new (and tested) factory method on PrepareStatement.

Improved documentation to also show that placeholders are ok in such
statements.

Results
-------
Users can now prepare a statement given in String format.

Change-Id: I90fc7b605b297b844615e131efe070eef62c6786
Reviewed-on: http://review.couchbase.org/45932
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>