JCBC-407: Allow DNS SRV bootstrap.
Motivation
----------
In order to allow users to centralize their seed nodes list in
DNS records, DNS SRV should be utilized to automatically fetch
those records.

Modifications
-------------
Not all DNS SRV informations are supported, weighting and priority
are not.

	- dnsSrvEnabled() must be set on the environment.
	- only one record is passed into the seed node list
	- this node is used as the dns srv lookup record
	- the fetched list is then used to bootstrap the cluster

Note that the fallback path is that the node passed in for DNS SRV
bootstrap will be used as the A record for further bootstrap.

Here is a sample set of records (replace _couchbases if you have
sslEnabled on the environment):

_couchbase._tcp.example.com.  3600  IN  SRV  0  0  0  node1.example.com.
_couchbase._tcp.example.com.  3600  IN  SRV  0  0  0  node2.example.com.
_couchbase._tcp.example.com.  3600  IN  SRV  0  0  0  node3.example.com.

Result
------
It is now possible to fetch a bootstrap list through DNS SRV.

Change-Id: I7168ff9a1341d2d1bd90c7b735e9217c4498610e
Reviewed-on: http://review.couchbase.org/45983
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>