JCBC-696: Expose the list of nodes directly on the BucketInfo.
Motivation
----------
Based on user feedback it was identified that direct access to the list
of nodes in the cluster/bucket makes it easier to consume it for
further processing at the user level.

Modifications
-------------
Iterate over the raw response, filter out the hostnames (without port)
and expose it directly as a getter. A test has been modified to make
sure it works as expected.

Result
------
Easier access to the list of nodes in that cluster.

Change-Id: Ia4538c8ace57a2ba426562d286f725cccc3c5e5b
Reviewed-on: http://review.couchbase.org/46030
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>