JCBC-539: Add NodeLocatorHelper to help mapping IDs to nodes.
Motivation
----------
In some cases it can be helpful to know to which node of the cluster
a specific document ID maps. In order to not expose lots of core-io
internals, a helper class is introduced which encapsulates the
functionality.

Modifications
-------------
The helper class subscribes itself to the configuration provider in
order to be always up-to-date on the latest configuration. It provides
both information on all nodes in the config, but also exposes methods
to explicitly map IDs to nodes.

Note that this class is stateful, since it subscribes to the changes
on the configuration provider.

Result
------
Easy and small surface area access to the underlying id => node
mapping.

Change-Id: I241c36eb652f06e333dc6eb27a5392a02b7d7b94
Reviewed-on: http://review.couchbase.org/46075
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>