JCBC-701: Error with DocumentDoesNotExistException on remove
Motivation
----------
Currently the catch-all CouchbaseException is thrown, which does
not provide the proper semantics needed.

Modifications
-------------
Throwing the proper exception makes it possible for the user to
apply fine grained control and retry logic as needed, so this
change adds a test to verify.

Result
------
Correct exception thrown on remove when the document is not found.

Change-Id: Icb2614e91d624ababc40a2f2794382001e91d1d1
Reviewed-on: http://review.couchbase.org/46310
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>