JCBC-702: Avoid flatMap for insert/upsert/replace.
Motivation
----------
Flatmap is more expensive (gc, syncing) and is not needed in this
case were a simple conversion is happening.

Modifications
-------------
flatMap is converted to a regular map, giving a slight performance
boost and better GC characteristics. In addition, it is now in
line with all the other methods in the bucket for similar
operations.

Result
------
Simpler implementation and slightly better performance and GC
behaviour.

Change-Id: I42d0fc2c120474aa187235da268feb1d5ded815e
Reviewed-on: http://review.couchbase.org/46324
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>