JCBC-704: Funnel all durability errors into a DurabilityException.
Motivation
----------
Currently the documentation states correctly that all durability errors which
violate the requirement are caught and raised in the DurabilityException.
This is currently only the case on one mutation method, all the others
need to follow suit.

Modifications
-------------
Properly catch and collect the errors on mutation and remove observe methods.

Also the BinaryTest has been renamed to KeyValueTest for clarity.

Result
------
Proper exception thrown as advertised in the docs.

Change-Id: I6afa1d5d9de6cf9e0295ad550624349f50d4d4ae
Reviewed-on: http://review.couchbase.org/46593
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>