JCBC-706: Make certain Document an JsonValue subclasses serializable.
Motivation
----------
While building spark integration, it was found that some objects need
to be serializable in order to be transferred over the network by
spark. We found that it also makes sense to make them serializable
directly in the java-client where the actual types support it.

Modifications
-------------
Serilization has been added to:

	- JsonObject
	- JsonArray
	- All Document* subtypes other than LegacyDocument and
	  BinaryDocument because their content types are not
	  serializable.

Tests have been added to prove its working.

Result
------
Transparent serialization for document types and json values.

Change-Id: I5a9a032376ec63b115500fa2916ac95cb6ac2ea7
Reviewed-on: http://review.couchbase.org/46679
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>