JCBC-707: Make sure to account for array offsets.
Motivation
----------
When the array is backing more than one ByteBuf the code needs to
make sure to only grab the slice which is actually belonging to
the specific ByteBuf.

Modifications
-------------
Calculate the offset + index range for the byte[] slice before
sending it over to jackson.

Result
------
More resilient extraction logic.

Change-Id: I46a80fce5584e179e34aea9a622b5139efecfdda
Reviewed-on: http://review.couchbase.org/46690
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>