JCBC-700: Set ObserveFuture isDone() properly.
Motivation
----------
When a ObserveFuture is returned (the underlying one which is always
returned when PersistTo or ReplicateTo overloads are used) the current
code always returns false on isDone, even when done.

Modifications
-------------
Overriding the set() method which is called by the callback to set
the proper status and count down the latch to also set the "done"
flag.

Adding a test case to verify its working.

Result
------
Proper behavior when calling isDone() on observe overloads.

Change-Id: I783d10753c1cbe7bbba4d3940922b78161c9023e
Reviewed-on: http://review.couchbase.org/46873
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>