JCBC-711: Prepare statements to be Serializable
Motivation
----------
In order to use N1QL capabilities in Spark, Query implementations must
be Serializable. This in turns necessitate the underlying statements to
be Serializable...

Modifications
-------------
Introduced an intermediary interface to explicitly distinguish between
the (non-serializable) DSL and other statements. Used this interface in
AbstractQuery to ensure all its fields are Serializable.

Each statement used in various Query implementations is now of the
SerializableStatement kind. When passing in a DSL, it is converted to a
Serializable alternative using RawStatement.

Result
------
Statements stored in Query implementations are now Serializable, paving
the way to fully Serializable queries.

Change-Id: I1ca76f8b8056bf2e7ef644358ea88691f3814eb9
Reviewed-on: http://review.couchbase.org/46934
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>