JCBC-711: Make Query and QueryParams serializable.
Motivation
----------
To allow N1QL queries to be run through spark and other components,
the queries need to be serializable. Also the params which are
attached to the queries need to be serializable.

Modifications
-------------
Query and QueryParams have been made serializable, tests added to
make sure it works properly.

Result
------
N1QL queries can be used through spark and other components that
require this functionality.

Change-Id: Ib73800d8eeff7ef4a9a2c0039cd0dca7d2199c00
Reviewed-on: http://review.couchbase.org/46935
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>