JCBC-717: expose N1QL signature
Motivation
----------
N1QL signature is relevant to some users reviewing N1QL. It is exposed
in the core since JVMCBC-153 and should be exposed as well in client.

Modifications
-------------
Added signature as an Observable<JsonObject> in AsyncQueryResult and as
a JsonObject in QueryResult.
Added an integration test for the signature.

This change depends on JVMCBC-153 in core.

Results
-------
Users can retrieve the signature section of the N1QL response as a
JsonObject.

Change-Id: I75d798b71a27b188b4bd5fbc273b01acb6eb8e37
Reviewed-on: http://review.couchbase.org/47230
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>