JCBC-723: touch unknown key throws DocumentDoesNotExistException
Motivation
----------
Unlike most similar operations, currently touch won't throw a specific
DocumentDoesNotExistException when the provided key is unknown, but a
more generic CouchbaseException with content "NOT_EXIST". It can be
expected that the correct and more specific exception be used.

Modifications
-------------
The async touch now throws the DocumentDoesNotExistException when the
provided key is unknown. This is thus true as well of the sync touch.
Tests and documentation has been updated accordingly.

Results
-------
Users receive a more meaningful exception when they try to touch a doc
that doesn't exist.

Change-Id: Iea0987230a305ebc2e881364ceeb5ee26425c742
Reviewed-on: http://review.couchbase.org/47946
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>