JCBC-735: N1QL signature is an Object
Motivation
----------
The N1QL response signature field was expected to be just a JSON sub
object but turns out to be any JSON entity.
Thus the QueryResponse signature method should return Object and not
JsonObject.

Modifications
-------------
QueryResult and AsyncQueryResult signature methods now deal with Object
instead of JsonObject.

The parsing of response's JSON to Object is done so that, when relevant,
a JsonObject or JsonArray is returned (not Map and List as in default
jackson unmarshalling).

For that, a new static method has been added to the JsonTranscoder. It
also ignores and skips leading whitespaces.

Test for the various signatures possibilities has been added.

Result
------
The fact that the N1QL's signature can be any JSON valid value is now
taken into account.

Change-Id: Ie5ca6165bcd6f3a10ab435553db06d80ae8f4d28
Reviewed-on: http://review.couchbase.org/48694
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>