JCBC-740: Expose View errors as Observable
Motivation
----------
A related core issue has been addressed (JVMCBC-176) and as
a result the view and spatial view results need to expose the
error in an observable instead of the plain JsonObject.

The reasoning behind that is the errors can also come after
rows have already been emitted. The core also moved the erorrs
from the info observable into a separate errors observable, which
makes it easier to grab errors but also needs slight changes
to the language binding in order to properly identify them.

Modifications
-------------
The signatures have been adapted from error JsonObject to
Observable<JsonObject>. Also, the response mapper has been
modified to align with those changes and most importantly
the retry handler has been adapted to do the maybeRetry
based on the error and not on the info portion.

Result
------
Correct behaviour when a view error is reported after the
rows have been emitted.

Change-Id: Ie984be68b00f5abd75fec3bf68a380bb963b741f
Reviewed-on: http://review.couchbase.org/48985
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>