JCBC-741: Always poll master on observe.
Motivation
----------
To consistently detect concurrent modifications of the
observed document, the client needs to always poll the
master cas state, since it can't be reliably detected
on the client.

So even when just ReplicateTo.* is used, the master
is polled and used for a cas crosscheck.

Modifications
-------------
The isMaster() check blocks have been removed which will
enable master polling all the time, leading to cas
mismatch detection.

Result
------
Consistent concurrent modification check behaviour,
independent if PersistTo is used or not.

Change-Id: Ia48cb827ff90a1d7511a42fd10fd058dd3dcb845
Reviewed-on: http://review.couchbase.org/49023
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>