JCBC-738: Avoid casting op in HttpFuture on timeout.
Motivation
----------
A bug was reported that the HttpFuture incorrectly casted the
operation to a net.spy one, which is unexpected. In fact, the
cast itself is unnecessary and can be removed.

Modifications
-------------
The timeout exception is now directly exposed to be in line with
the other timeout case on the codepath and as a result is not
subject to a classcast exception anymore as well as consistently
throws a TimeoutException in the timeout case. In fact, the
method directly allows a TimeoutException to be thrown as a
checked exception.

Result
------
No classcast exceptions and correct exception type thrown on all
timeout codepaths.

Change-Id: I4e0cda493a4cba961c42cef3b24fdea49e18fb81
Reviewed-on: http://review.couchbase.org/49020
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>