JCBC-730: Implement exists() functionality.
Motivation
----------
The exists() method, internally based on the observe command, can
be used to see if a document exists on the server without having
to load the full document.

Modifications
-------------
This changeset implements the exists() method based on the observe
command, by sending a request to the active partition and checking
if it is either persisted or just stored in memory.

Result
------
exists functionality is now available on the bucket api.

Change-Id: Ib31fc8583f5fb8831297e569f3eb29c580e4481f
Reviewed-on: http://review.couchbase.org/49440
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>