JCBC-732: Extend Expression class.
Motivation
----------
As part of the DSL overhaul for 4.0, the Expression class needs to
be extended.

Modifications
-------------
The Expression class has been overhauled, tests added and lots of overloads
have been put in place so that most of the time, x() can be avoided.

Result
------
More tests and less clutter when working with expressions.

Change-Id: I65749a0e59ba8be8ec2acc121a9e45876b9892bb
Reviewed-on: http://review.couchbase.org/49449
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>