JCBC-744: Add support for @Field alias
Motivation
----------
It should be possible to have a different name in the JSON field
than the field name in the object.

Modifications
-------------
The user can now supply an explicit @Field("alias") which will be
used as the JSON field key instead of the actual object field name.

Result
------
More flexible field mapping.

Change-Id: If783ea54a8fa3334f86c15864c5524a7f9d7934e
Reviewed-on: http://review.couchbase.org/49859
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>