JCBC-744: Add metadata caching.
Motivation
----------
In order to avoid the overhead of re-generating the metadata all
the time, a simple CHM has been added to cache the metadata once
generated.

Modifications
-------------
A ConcurrentHashMap stores the already generated metadata. Note that
there is currently no way of "expiring" items, but the number of
entities during runtime remain static and relatively small, most
of the time (since the expectation is that they are not generated
new at runtime).

Result
------
Better performance by avoiding metadata generation overhead all
the time.

Change-Id: I0a96f8e318118036f6ddbf49dfb7d886e5c8d976
Reviewed-on: http://review.couchbase.org/49890
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>