JCBC-744: Enforce String @Id type.
Motivation
----------
Since the document ID needs to be a string, the code should check
that early as part of the entity mapping.

Modifications
-------------
A type check has been added on the type and now reports an exception
when the String assertion does not hold.

Result
------
Well defined boundary on what is supported for the @Id types.

Change-Id: I6e4c120379253f60160cf6819ae25c34c024741e
Reviewed-on: http://review.couchbase.org/49891
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>