JCBC-770: Memcached buckets fail to pick up new configuration.
Motivation
----------
For memcached buckets, the client was able to pick up an initial
configuration, but failed to properly establish a streaming
connection to get subsequent configuration updates.

This resulted in the problem that when memcached nodes are
added, it is not picked up properly and since the ketama
algorithm doesn't make it fail it could only be observed
by looking in the UI. Also, node removals are not
properly picked up.

Modifications
-------------
The root cause of the problem is that the HTTP streaming
connection is not properly attached to one of the servers,
and the reason for that is that the correct bootstrap provider
type was not set and therefore an assertion failed.

Result
------
After setting the bootstrap provider before monitoring the
bucket through a streaming connection, new configs are now
properly picked up (both node add and removals) and delivered
to the subscribers.

Change-Id: I4565b1d7f6c28773736be0a18a020029f0b7964e
Reviewed-on: http://review.couchbase.org/51153
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>