JCBC-772: query(Statement) needs to check for a QueryPlan
Motivation
----------
the query(Statement) is intended as a shortcut, but also needs
to take into account that a QueryPlan might be passed in.

Modifications
-------------
The method now checks if the input statement is a query plan and
if so wraps it into a Query.prepared() instead of a Query.simple().

code like bucket.query(QueryPlan) is now possible.

Result
------
Easier way to query with a query plan.

Change-Id: I9d86ed082d83746b53b3b248f2dfbc3755bd5404
Reviewed-on: http://review.couchbase.org/51210
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>