JCBC-774 join should accept Expressions in N1QL DSL
Motivation
----------
Unlike most other cases (like from()), the operations on the LetPath
don't accept Expressions, only Strings. This is counterintuitive when
one wants to escape a namespace in a JOIN for example.

Modifications
-------------
Added Expression overloads in the LetPath (joins, nests, unnests).
Added corresponding tests.

Results
-------
Users of the DSL can now do a JOIN on an Expression, allowing them to
escape the namespace for example.

Change-Id: I35012077e362861188dfe2478347356f2c6365d2
Reviewed-on: http://review.couchbase.org/51438
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>