JCBC-780: add sortCount to QueryMetrics
Motivation
----------
The Query team introduced a new metric that shows the number of results
before it is trimmed down (eg. by a LIMIT clause). This should be also
exposed as a QueryMetrics method.

Modifications
-------------
Added sortCount() method to QueryMetrics.
Modified unit tests to account for it.

Results
-------
Users can now see this metric directly, not having to fall back to the
JsonObject raw version of the metric representation.

Note that if the metric is not returned by the engine, this method
returns 0.

Change-Id: I3cb309f65dd60a2c0899972e76d753df6fdc30d6
Reviewed-on: http://review.couchbase.org/51950
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>