JCBC-789: Remove AtomicReferences from Blocking#blockForSingle
Motivation
----------
While profiling SDK code, it was found that there is a portion
of AtomicReferences allocated. It turns out that they can be
replaced with volatiles only, since they only use get/set
semantics.

Modifications
-------------
The subscriber has been refactored out of the helper method since
volatiles are not allowed inside a method. The AtomicReferences
have been replaced with plain volatile calls.

Result
------
Less allocations on the hot code paths, since all the features
of an AtomicReference are not even needed.

Change-Id: I6583a6cee10c9d9ec5f73c15b03273295a7ef5e3
Reviewed-on: http://review.couchbase.org/52422
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>