JCBC-767: add use index hint to query DSL
Motivation
----------
The Query team introduced a new clause to hint at which index(es) to use
in a query. This comes in the form of a USE INDEX clause after the FROM
clause.

Modifications
-------------
Added HintPath and HintIndexElement to represent this new clause.
Introduced a new IndexReference class that wraps an escaped index name
and an optional USING GSI|VIEW suffix.
Added unit tests for the HintPath.

Results
-------
Users can now hint at which index(es) to use. New tests pass.

Change-Id: I99dcd4d90730f605a48951f8088cffddc46e0e89
Reviewed-on: http://review.couchbase.org/52590
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>