JCBC-790: Ignore partial replica read failures on ReplicaMode.ALL
Motivation
----------
The accompanying change in core-io now quickly fails if the replica
is not available. This change silents those errors, leading to
the fact that if the replica is not available, only a subset
(the responding) replicas is returned.

Modifications
-------------
The replica read code has been refactored into its own class for
clarity and better future testability. The code is restructured
so that individual filtering and error handling happens on
a per-operation basis, errors are silenced and logged.

API docs have been updated to reflect this new behaviour.

Result
------
More resilient getFromReplica calls when replicas are not
available.

Change-Id: I00ef787d57a4acb76b344e66476730769ef08318
Reviewed-on: http://review.couchbase.org/52702
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-on: http://review.couchbase.org/52705