JCBC-652: Support DesignDocument Options
Motivation
----------
The view engine currently supports 2 options which can be overridden
on a per desing-document basis: updateMinChanges and
replicaUpdateMinChanges. They need to be exposed on the DesignDocument
API so they can be set.

Modifications
-------------
The DesignDocument has been extended to optionally support setting
those two options and will send them as part of the HTTP request
if needed.

Result
------
The options can now be configured by the user when a design document
is created or modified.

Change-Id: I47215116d849777276312d7e0ab021fba713ff9a
Reviewed-on: http://review.couchbase.org/52828
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>