JCBC-810: Cache getFromReplica to avoid eager unsubscription leaks.
Motivation
----------
A common pattern to execute is getFromReplica(ALL) and then use methods
like first() or take(n) to get specific objects out of the observable.

A problem has been identified with eager unsubscriptions and object
leakage (the contained ByteBufs are not properly freed on the unsubscribed
observable).

Modifications
-------------
By using the cache() operator, the emitted values are cached and a potential
unsubscribe is disconnected. Since a maximum of 4 values can be emitted, this
should not be a problem.

The ReplicaMode has been extended to provide how many nodes might be affected
so that a size hint can be provided to cache() and keep the memory overhead
as low as possible.

Result
------
Using first() or take(n) methods on the getFromReplica() method work without
issues and leaks.

Change-Id: Ieae5c08f1edd41745c9c2d64db657cdc5208fdb7
Reviewed-on: http://review.couchbase.org/53218
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>