JCBC-782: Make CouchbaseAsyncBucket cold.
Motivation
----------
Since all the calls against the core layer are considered hot,
this brings along some limitations which are not obvious. The
main issues we ran into in the field are:

	- Retry logic can't work because the retried subject
	  is the same as the failed one.
	- Fallback code is eagerly evaluated, leading to
	  inconsistent and surprising results

By wrapping the hot observables into a defer method, they
work like cold observables.

Modifications
-------------
This changeset only cares about the CouchbaseAsyncBucket
to not make the changeset larger than it should be.

Result
------
All methods on the CAB are now cold.

Change-Id: Iecfed8697418932d746b32660adc1b2765b44eb3
Reviewed-on: http://review.couchbase.org/53228
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>