JCBC-782: Make Other Async APIs cold.
Motivation
----------
Since all the calls against the core layer are considered hot,
this brings along some limitations which are not obvious. The
main issues we ran into in the field are:

    - Retry logic can't work because the retried subject
      is the same as the failed one.
    - Fallback code is eagerly evaluated, leading to
      inconsistent and surprising results

By wrapping the hot observables into a defer method, they
work like cold observables.

Modifications
-------------
This changeset moves the rest of the async API to cold.

Result
------
All other methods on the async API are now cold.

Change-Id: I02ba50d938595ba21810b647a02ad146b164ad82
Reviewed-on: http://review.couchbase.org/53232
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>